+++ test/instabug_flutter_test.dart
@@ -89,8 +89,8 @@ test('startWithToken:invocationEvents: Test', () async {
   });
   
   test('setLocale:', () async {
-    Instabug.setLocale(Locale.German);
-    final List<dynamic> args = <dynamic>[Locale.German.toString()];
+    Instabug.setLocale(Locale.german);
+    final List<dynamic> args = <dynamic>[Locale.german.toString()];
     expect(log, <Matcher>[
       isMethodCall('setLocale:',
         arguments: args,
@@ -254,8 +254,8 @@ test('startWithToken:invocationEvents: Test', () async {
 
   
   test('invokeWithMode:options: Test', () async {
-    BugReporting.invokeWithMode(InvocationMode.BUG, [InvocationOption.COMMENT_FIELD_REQUIRED]);
-    final List<dynamic> args = <dynamic>[InvocationMode.BUG.toString(), <String>[InvocationOption.COMMENT_FIELD_REQUIRED.toString()]];
+    BugReporting.invokeWithMode(InvocationMode.bug, [InvocationOption.commentFieldRequired]);
+    final List<dynamic> args = <dynamic>[InvocationMode.bug.toString(), <String>[InvocationOption.commentFieldRequired.toString()]];
     expect(log, <Matcher>[
       isMethodCall('invokeWithMode:options:',
         arguments: args,
@@ -275,7 +275,7 @@ test('startWithToken:invocationEvents: Test', () async {
 
   test('test setValueForStringWithKey should be called with two arguments', () async {
     const String value = 'Some key';
-    const IBGCustomTextPlaceHolderKey key = IBGCustomTextPlaceHolderKey.SHAKE_HINT;
+    const IBGCustomTextPlaceHolderKey key = IBGCustomTextPlaceHolderKey.shakeHint;
     Instabug.setValueForStringWithKey(value, key);
     final List<dynamic> args = <dynamic>[value, key.toString()];
     expect(log, <Matcher>[
@@ -390,7 +390,65 @@ test('startWithToken:invocationEvents: Test', () async {
     ]);
   });
 
+  test('setInvocationEvents Test', () async {
+     BugReporting.setInvocationEvents(<InvocationEvent>[InvocationEvent.floatingButton]);
+    final List<dynamic> args = <dynamic>[<String>[InvocationEvent.floatingButton.toString()]];
+    expect(log, <Matcher>[
+      isMethodCall('setInvocationEvents:',
+        arguments: args,
+      )
+    ]);
+  });
 
+  test('setEnabledAttachmentTypes:extraScreenShot:galleryImage:screenRecording: Test', () async {
+    BugReporting.setEnabledAttachmentTypes(false, false, false, false);
+    final List<dynamic> args = <dynamic>[false, false, false, false];
+    expect(log, <Matcher>[
+      isMethodCall('setEnabledAttachmentTypes:extraScreenShot:galleryImage:screenRecording:',
+        arguments: args,
+      )
+    ]);
+  });
+
+ test('setInvocationEvents Test', () async {
+    BugReporting.setReportTypes(<ReportType>[ReportType.feedback]);
+    final List<dynamic> args = <dynamic>[<String>[ReportType.feedback.toString()]];
+    expect(log, <Matcher>[
+      isMethodCall('setReportTypes:',
+        arguments: args,
+      )
+    ]);
+  });
+
+  test('setInvocationEvents Test', () async {
+    BugReporting.setExtendedBugReportMode(ExtendedBugReportMode.enabledWithOptionalFields);
+    final List<dynamic> args = <dynamic>[ExtendedBugReportMode.enabledWithOptionalFields.toString()];
+    expect(log, <Matcher>[
+      isMethodCall('setExtendedBugReportMode:',
+        arguments: args,
+      )
+    ]);
+  });
+
+  test('setInvocationOptions Test', () async {
+    BugReporting.setInvocationOptions(<InvocationOption>[InvocationOption.emailFieldHidden]);
+    final List<dynamic> args = <dynamic>[<String>[InvocationOption.emailFieldHidden.toString()]];
+    expect(log, <Matcher>[
+      isMethodCall('setInvocationOptions:',
+        arguments: args,
+      )
+    ]);
+  });
+
+   test('showBugReportingWithReportTypeAndOptions:options Test', () async {
+   BugReporting.showWithOptions(ReportType.bug, <InvocationOption>[InvocationOption.emailFieldHidden]);
+    final List<dynamic> args = <dynamic>[ReportType.bug.toString(), <String>[InvocationOption.emailFieldHidden.toString()]];
+    expect(log, <Matcher>[
+      isMethodCall('showBugReportingWithReportTypeAndOptions:options:',
+        arguments: args,
+      )
+    ]);
+  });
 
 }